requirements.txt
streamlit
langchain
langchain-google-genai
google-generativeai
pillow
python-dotenv
